Fixes #14. Also fixed a comment typo.
Removed ambiguous characters when viewed as either upper or lower case, in most fonts, and in lower quality printing (i.e., plastic gift cards).

I did a lot of research on unambiguous character sets and came up with `CFHMNPRSTWXY347`.

Coupon codes are not case sensitive in WooCommerce so if a character in one case had to be removed then the other case had to be removed as well. Store owners may choose to display coupon codes for customers in upper or lower case so it should be unambiguous in each.

Ambiguous sets removed:
1, i, I, l, L
o, 0, O, D, Q
2, z, Z
5, S, s
6, b, G
8, B
9, g, q
U, V, u, v
K, k (can be confused with X, but X is pretty clear so left in)
a, e, i, o, u (No vowels to avoid offensive words being formed)

There are others that could be removed but the character set is already pretty small. For example, R, N, and M. In lower case, rn or nn can look like m. The same could be said of vv and w. If codes are displayed by the store owner in upper case, like this plugin generates and WooCommerce displays then this is not a problem.

With only this 15 character set, it becomes more important to generate longer codes or there will likely be birthday paradox type collisions causing duplicate codes. According to this [Stack Overflow discussion about collisions](https://stackoverflow.com/questions/7591117/what-is-the-probability-of-collision-with-a-6-digit-random-alphanumeric-code) and playing with the calculations it links to, an [8 character code](https://www.wolframalpha.com/input/?i=1%2F2%2Bsqrt%281%2F4-2*%2815%5E8%29*ln%280.5%29%29) is likely to have a 1 in 60,000 chance of collisions. A [12 character code](https://www.wolframalpha.com/input/?i=1%2F2%2Bsqrt%281%2F4-2*%2815%5E12%29*ln%280.5%29%29) is better at 1 in about 13 million. It takes a [16 character code](https://www.wolframalpha.com/input/?i=1%2F2%2Bsqrt%281%2F4-2*%2815%5E16%29*ln%280.5%29%29) to get into the low billions.

So is this too aggressive in trimming the character set?

I'm not seeing any check for an existing code before writing it. Doing so would slow things down a bit but would mitigate the occasional collision if it occurred.